Improve Hugo-Server.ps1 script

+++ Hugo-Server.ps1
@@ -1,3 +1,8 @@
-Remove-Item -Recurse -Force .\resources
-Remove-Item -Recurse -Force .\public
-hugo server -D --disableFastRender
+if (Test-Path .\resources) {
+    Remove-Item -Recurse -Force .\resources
+}
+if (Test-Path .\public) {
+    Remove-Item -Recurse -Force .\public
+}
+
+Start-Process hugo -ArgumentList "server -D --disableFastRender" -NoNewWindow -Wait
